refactor: Replace hardcoded enum validation arrays with actual enum constants
- Import RiskCategory and RiskStatus from @/types/rcsa.types
- Define TreatmentStrategy enum locally since not exported from rcsa.types
- Replace hardcoded validation arrays with Object.values(EnumName)
- Update imports in BasicInfoStep and DetailsStep to use correct sources
- Ensures type safety and prevents mismatch between validation and actual enum values
