+++ src/components/risks/NewRiskFlow/steps/DetailsStep.tsx
@@ -8,8 +8,8 @@ import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import { Textarea } from '@/components/ui/textarea';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
-import { useRiskFlow } from '../RiskFlowContext';
-import { RiskStatus, TreatmentStrategy } from '@/types';
+import { useRiskFlow, TreatmentStrategy } from '../RiskFlowContext';
+import { RiskStatus } from '@/types/rcsa.types';
 import { cn } from '@/lib/utils';
 import { format } from 'date-fns';
